progress: split progress.go into command.go
The progress.go file is overloaded with functionality, so let's split it into
command.go and progress.go. The command.go will handle the command-line
interface, while progress.go will focus on the core progress tracking logic.
